Split "Enums" into "DataEnums" and "BattleEnums"
